Telegram_mts/requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
config==0.5.1
et-xmlfile==1.1.0
frozenlist==1.3.1
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
multidict==6.0.2
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.1
openpyxl==3.0.10
pandas==1.4.3
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
telebot==0.0.4
telegram==0.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
yarl==1.8.1
